Add support for Öppet Arkiv
This commit adds support for the new SVTPlay offering called "Öppet
Arkiv" which is an open archive with a lot of old stuff from swedish
television history.

The site is almost identical to SVTPlay so it reuses a lot already with
only minor modifications.

Support for this is also enabled in the SVTPlay plugin.