Fix for SVTPlay on Chrome and Android - Remove subs for HLS
Many of the urls for HLS coming from SVTPlay looks like this:

 http://svtplay10p-f.akamaihd.net/i/se/open/20140502/1343328-009A/MASTARNAS_MASTA-009A-3c383763a753ed68_,900,348,564,1680,2800,.mp4.csmil/master.m3u8?cc1=name=Svenska~default=yes~forced=no~uri=http://media.svt.se/download/mcc/wp3/undertexter-wsrt/1343328/1343328-009A/C(sv)/index.m3u8~lang=sv

The parameters to the url is causing many players to fail to play.
Safari plays this without problem. Including showing the subs.
Chrome plays this if I replace index.m3u8~ with index.m3u8?~ But
still doesn't show the subs.
Android doesn't play this regardless of what I do.

The solution is to remove all parameters behind the url.

The drawback is that Safari, the only player to play this without
issues will no longer get the subs.